+++ src/main/java/com/github/doragonUni/finalreality/model/character/AbstractCharacter.java
@@ -10,26 +10,63 @@ import com.github.doragonUni.finalreality.model.weapon.IWeapon;
 import org.jetbrains.annotations.NotNull;
 
 /**
- * An abstract class that holds the common behaviour of all the characters in the game.
+ * An abstract class that holds the common behaviour of all the PLAYER characters in the game.
  *
  * @author Ignacio Slater Muñoz.
  * @author <Your name>
  */
 public abstract class AbstractCharacter implements ICharacter {
-
+  private ScheduledExecutorService scheduledExecutor;
   protected IWeapon equippedWeapon = null;
+
+
   protected final BlockingQueue<ICharacter> turnsQueue;
   private final String name;
-  //private Weapon equippedWeapon = null; down to playercharacter
-  private ScheduledExecutorService scheduledExecutor;
+  private int hp;
+  private int defense;
+
 
   protected AbstractCharacter(@NotNull BlockingQueue<ICharacter> turnsQueue,
-                              @NotNull String name) {
+                              @NotNull String name,
+                              int hp,
+                              int defense) {
     this.turnsQueue = turnsQueue;
     this.name = name;
+    this.hp = hp;
+    this.defense = defense;
   }
 
-  //wait durn divided into enemy and playerChARACTER
+
+
+  /**
+   * gets the name (String) of the Character
+   */
+  @Override
+  public String getName() {
+    return this.name;
+  }
+
+  /**
+   * gets the name Hp of the Character
+   */
+  public int getHp(){
+    return this.hp;
+  }
+
+  /**
+   * gets the defense of the Character
+   */
+  public int getDef(){
+    return this.defense;
+  }
+
+
+  /**
+   * gets the Weapon equipped by the Character
+   */
+  public IWeapon getEquippedWeapon() {
+    return this.equippedWeapon;
+  }
 
   /**
    * Adds this character to the turns queue.
@@ -39,30 +76,15 @@ public abstract class AbstractCharacter implements ICharacter {
     scheduledExecutor.shutdown();
   }
 
-  @Override
-  public String getName() {
-    return name;
-  }
-
-  public IWeapon getEquippedWeapon() {
-    return this.equippedWeapon;
-  }
-
+  /**
+   * function that wait the turn of the round
+   */
   @Override
   public void waitTurn() {
     scheduledExecutor = Executors.newSingleThreadScheduledExecutor();
-
     scheduledExecutor
             .schedule(this::addToQueue, equippedWeapon.getWeight() / 10, TimeUnit.SECONDS);
   }
-
-
-
-
-    //void equipWeapon down to player character
-  // weapon getEquippedWeapon DOWN TO PLAYERCHARACTER
-
-
 
 }