+++ src/main/java/com/github/doragonUni/finalreality/model/character/Enemy.java
@@ -17,32 +17,67 @@ import org.jetbrains.annotations.NotNull;
 public class Enemy implements ICharacter {
 
   protected final BlockingQueue<ICharacter> turnsQueue;
-  protected final String name;
   private ScheduledExecutorService scheduledExecutor;
+  protected final String name;
   private final int weight;
+  private final int hp;
+  private final int defense;
 
   /**
    * Creates a new enemy with a name, a weight and the queue with the characters ready to
    * play.
    */
-  public Enemy(@NotNull final String name, final int weight,
-               @NotNull final BlockingQueue<ICharacter> turnsQueue) {
+  public Enemy(@NotNull final String name, @NotNull final BlockingQueue<ICharacter> turnsQueue,
+               final int weight, final int hp, final int defense) {
+    /**
+     * Creates a new Enemy.
+     *
+     * @param name       the character's name
+     * @param weight      enemy's weight
+     * @param turnsQueue     the queue with the characters waiting for their turn
+     * @param hp        this character's health points
+     * @param defense    this character defense points
+     *
+     */
     this.turnsQueue = turnsQueue;
     this.name = name;
     this.weight = weight;
+    this.hp = hp;
+    this.defense = defense;
 
   }
 
-  /**
-   * Returns the weight of this enemy.
-   */
 
+
+  /**
+   * gets the Enemy Health Points
+   */
+  @Override
+  public int getHp() {
+    return this.hp;
+  }
+
+
+  /**
+   * gets the Enemy Defense
+   */
+  @Override
+  public int getDef() {
+    return this.defense;
+  }
+
+  /**
+   *Adds an Enemy to the Queue
+   */
   @Override
   public void addToQueue() {
     turnsQueue.add(this);
     scheduledExecutor.shutdown();
   }
 
+  /**
+   * @see ICharacter
+   */
   @Override
   public void waitTurn() {
     scheduledExecutor = Executors.newSingleThreadScheduledExecutor();
@@ -51,11 +86,18 @@ public class Enemy implements ICharacter {
             .schedule(this::addToQueue, enemy.getWeight() / 10, TimeUnit.SECONDS);
   }
 
+
+  /**
+   * gets the Enemy name
+   */
   @Override
   public String getName() {
     return this.name;
   }
 
+  /**
+   * gets the Enemy Weight
+   */
   public int getWeight() {
     return weight;
   }
@@ -72,7 +114,9 @@ public class Enemy implements ICharacter {
     }
     final Enemy enemy = (Enemy) o;
     return getWeight() == enemy.getWeight() &&
-            getName().equals(enemy.getName());
+            getName().equals(enemy.getName()) &&
+            getHp() == enemy.getHp() &&
+            getDef() == enemy.getDef();
   }
 
   @Override