+++ src/main/java/com/github/doragonUni/finalreality/model/character/player/Knight.java
@@ -15,28 +15,38 @@ import java.util.concurrent.TimeUnit;
 
 public class Knight extends AbstractCharacter {
     /**
-     * Creates a new character.
+     * Creates a new Knight.
      *
      * @param name       the character's name
-     * @param turnsQueue
+     * @param turnsQueue     the queue with the characters waiting for their turn
+     * @param hp        this character's health points
+     * @param defense    this character defense points
      */
 
 
 
 
 
-    public Knight(@NotNull String name, @NotNull BlockingQueue<ICharacter> turnsQueue) {
-        super(turnsQueue, name);
+    public Knight(@NotNull String name, @NotNull BlockingQueue<ICharacter> turnsQueue, int hp, int defense) {
+        super(turnsQueue, name, hp, defense);
+
     }
 
+    /**
+     * equip an Axe to this Knight
+     */
     public void equipAxe(Axe axe) {
         this.equippedWeapon = axe;
     }
-
+    /**
+     * equip a Sword to this Knight
+     */
     public void equipSword(Sword sword) {
         this.equippedWeapon = sword;
     }
-
+    /**
+     * equip a Knife to this Knight
+     */
     public void equipKnife(Knife knife) {
         this.equippedWeapon = knife;
     }
@@ -50,7 +60,9 @@ public class Knight extends AbstractCharacter {
             return false;
         }
         final Knight that = (Knight) o;
-        return getName().equals(that.getName());
+        return getName().equals(that.getName()) &&
+                getHp() == that.getHp() &&
+                getDef() == that.getDef();
     }
 
     @Override