+++ src/main/java/com/github/doragonUni/finalreality/model/character/player/Thief.java
@@ -17,28 +17,39 @@ public class Thief extends AbstractCharacter {
 
 
     /**
-     * Creates a new character.
+     * Creates a new Thief.
      *
      * @param name           the character's name
      * @param turnsQueue     the queue with the characters waiting for their turn
-     *
+     * @param hp        this character's health points
+     * @param defense    this character defense points
      */
 
 
-    public Thief(@NotNull String name, @NotNull BlockingQueue<ICharacter> turnsQueue) {
-        super(turnsQueue, name);
+    public Thief(@NotNull String name, @NotNull BlockingQueue<ICharacter> turnsQueue, int hp, int defense) {
+        super(turnsQueue, name, hp, defense);
     }
 
-
+    /**
+     * equip a Sword to this Thief
+     */
     public void equipSword(Sword sword) {
         this.equippedWeapon = sword;
     }
 
+    /**
+     * equip a Bow to this Thief
+     */
     public void equipBow(Bow bow) {
         this.equippedWeapon = bow;
     }
 
-    public void equipKnife(Knife knife) { this.equippedWeapon = knife; }
+    /**
+     * equip a Knife to this Thief
+     */
+    public void equipKnife(Knife knife) {
+        this.equippedWeapon = knife;
+    }
 
 
     @Override
@@ -50,7 +61,9 @@ public class Thief extends AbstractCharacter {
             return false;
         }
         final Thief that = (Thief) o;
-        return getName().equals(that.getName());
+        return getName().equals(that.getName()) &&
+                getHp() == that.getHp() &&
+                getDef() == that.getDef();
     }
 
     @Override