+++ src/main/java/com/github/doragonUni/finalreality/model/character/player/WhiteMage.java
@@ -13,24 +13,35 @@ import java.util.concurrent.TimeUnit;
 
 public class WhiteMage  extends AbstractCharacter {
     /**
-     * Creates a new character.
+     * Creates a new WhiteMage.
      *
      * @param name       the character's name
-     * @param turnsQueue
-     * @param mana
+     * @param turnsQueue the queue with the characters waiting for their turn
+     * @param hp         this character's health points
+     * @param defense    this character defense points
+     * @param mana       mana value for this mage
      */
 
     private int mana;
 
-    public WhiteMage(@NotNull String name, @NotNull BlockingQueue<ICharacter> turnsQueue, int mana) {
-        super(turnsQueue, name);
+    public WhiteMage(@NotNull String name, @NotNull BlockingQueue<ICharacter> turnsQueue, int hp, int defense,
+                     int mana) {
+        super(turnsQueue, name, hp, defense);
         this.mana = mana;
     }
 
+
+    /**
+     * equip a Staff to this WhiteMage
+     */
     public void equipStaff(Staff staff) {
         this.equippedWeapon = staff;
     }
 
+
+    /**
+     * get the Mana value of this WhiteMage
+     */
     public int getMana(){
         return this.mana;
     }
@@ -46,8 +57,9 @@ public class WhiteMage  extends AbstractCharacter {
         }
         final WhiteMage that = (WhiteMage) o;
         return getName().equals(that.getName()) &&
+                getHp() == that.getHp() &&
+                getDef() == that.getDef() &&
                 getMana() == that.getMana();
-
     }
     @Override
     public int hashCode() {