+++ src/test/java/com/github/cc3002/finalreality/model/character/ThiefTest.java
@@ -12,40 +12,79 @@ import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.BlockingQueue;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 public class ThiefTest {
 
 
     protected BlockingQueue<ICharacter> turns;
     private static final String name = "Thief";
-    protected Sword sword = new Sword("sword", 10, 5 );
+    private int hp = 670;
+    private int def = 100;
+
+    protected Sword sword = new Sword("sword", 5, 50 );
     protected Bow bow = new Bow("bow", 10 , 5 );
-    protected Knife knife = new Knife( "knife",10, 5);
+    protected Knife knife = new Knife( "knife",15, 15);
 
     private Thief testThief;
-
+    private Thief testThief1;
+    private Thief testThief2;
+    private Thief testThief3;
+    private Knight fakeThief;
+    /**
+     * SETUP FOR TESTING
+     */
     @BeforeEach
     void setUp(){
-        testThief = new Thief(name, turns);
+        testThief = new Thief(name, turns, hp, def);
+        testThief1 = new Thief("T1", turns, hp, def);
+        testThief2 = new Thief(name, turns, 1003, def);
+        testThief3 = new Thief(name, turns, hp, 10000000);
+        fakeThief = new Knight("dishonor", turns, hp, def);
     }
-
+    /**
+     * Constructor TEST
+     */
     @Test
     void constructorTest(){
-        var expectedEngineer = new Thief(name, turns);
-        assertEquals(expectedEngineer, testThief);
-        assertEquals(expectedEngineer.hashCode(), testThief.hashCode());
-        assertNull(testThief.getEquippedWeapon());
+        var expectedThief = new Thief(name, turns, hp, def);
+        assertEquals(expectedThief, testThief);
+        assertEquals(expectedThief.hashCode(), testThief.hashCode());
+        assertNotEquals(testThief.hashCode(), fakeThief.hashCode());
+        assertEquals(testThief, testThief);
+        assertNotEquals(testThief, fakeThief);
 
+        //branches
+
+        assertNotEquals(expectedThief, testThief1);
+        assertNotEquals(expectedThief, testThief2);
+        assertNotEquals(expectedThief, testThief3);
+
+
+    }
+    /**
+     * TEST FOR THE GETTERS
+     */
+    @Test
+    void getTest(){
+        assertEquals(name, testThief.getName());
+        assertEquals(hp, testThief.getHp());
+        assertEquals(def, testThief.getDef());
+    }
+
+    /**
+     * TEST FOR EQUIPPING WEAPONS
+     */
+    @Test
+    void equipTest(){
+        assertNull(testThief.getEquippedWeapon());
         testThief.equipBow(bow);
         assertEquals(bow, testThief.getEquippedWeapon());
         testThief.equipKnife(knife);
         assertEquals(knife, testThief.getEquippedWeapon());
         testThief.equipSword(sword);
         assertEquals(sword, testThief.getEquippedWeapon());
-
-        assertEquals(name, testThief.getName());
 
     }